refactor: restructure DataTable component and split into smaller subcomponents
